Add instance_id field in Process Stats response
- Display the Diego instance identifier as part of the Process Stats
response. This id is unique for each actual LRP instance and can be used
as a substitute for the integer index id.
- Eirini does not support this in their API at this time, so the OPI
variant of the instances stats reporter returns an empty string for this
field. In the future, if Eirini is enhanced to provide this we can
update that client.
- Issue: https://github.com/cloudfoundry/cloud_controller_ng/issues/2819

Authored-by: Tim Downey <tdowney@vmware.com>